+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
 - Use internal coordinator communication for cursors if specified coordinator was not found on endpoint list
 - Add support for Overwrite Mode (ArangoDB 3.7)
+- Add support for Schema Collection options (ArangoDB 3.7)
 
 ## [1.0.0](https://github.com/arangodb/go-driver/tree/1.0.0) (N/A)
 - Enable proper CHANGELOG and versioning